Cleanup, PEP-8 compliance, etc...
