+++ impala/_rpc/hiveserver2.py
@@ -35,9 +35,9 @@ from decimal import Decimal
 from six.moves import range
 
 from impala.error import HiveServer2Error
-from impala._rpc.thrift_util import (get_socket, get_transport)
+from impala._thrift_api import (get_socket, get_transport, TTransportException,
+                                TBinaryProtocol)
 from impala._thrift_api.hiveserver2 import (
-    TTransportException, TBinaryProtocol,
     TOpenSessionReq, TFetchResultsReq, TCloseSessionReq, TExecuteStatementReq,
     TGetInfoReq, TGetInfoType, TTypeId, TFetchOrientation,
     TGetResultSetMetadataReq, TStatusCode, TGetColumnsReq, TGetSchemasReq,
@@ -155,7 +155,7 @@ def connect_to_impala(host, port, timeout=45, use_ssl=False, ca_cert=None,
     elif six.PY3:
         sock.set_timeout(timeout * 1000.)
     transport = get_transport(sock, host, kerberos_service_name, auth_mechanism,
-        user, password)
+                              user, password)
     transport.open()
     protocol = TBinaryProtocol(transport)
     if six.PY2:
